backend/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
python-multipart==0.0.20
groq==0.14.0
python-dotenv==1.0.1
